pyproject.toml
[tool.poetry]
name = "han"
version = "0.1.0"
description = ""
authors = ["Churnobyl <tjdcjfals@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.6"
django = "^4.2.2"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.2"
psycopg2 = "^2.9.6"
django-cors-headers = "^4.0.0"
python-dotenv = "^1.0.0"
pillow = "^9.5.0"
djangorestframework-simplejwt = "^5.2.2"
django-allauth = "0.51.0"
requests = "^2.31.0"
bs4 = "^0.0.1"
selenium = "^4.9.1"
channels = {extras = ["daphne"], version = "^4.0.0"}
redis = "^4.5.5"
channels-redis = "^4.1.0"
django-redis = "^5.2.0"
dj-rest-auth = "^4.0.1"
drf-social-oauth2 = "^2.1.3"
django-redis-sessions = "^0.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
